Moved upload folder to routes, created as documents are uploaded.
